+++ myokit/formats/cuda/template/kernel.cu
@@ -71,11 +71,11 @@ initial_values = model.initial_values(True)
 
 # Process bindings, remove unsupported bindings, get map of bound variables to
 # internal names
-bound_variables = model.prepare_bindings({
+bound_variables = myokit._prepare_bindings(model, {
     #'time' : 'time',
     #'pace' : 'pace',
     'diffusion_current' : 'I_diff',
-    })
+})
 
 # Get equations
 equations = model.solvable_order()